hardware: push multiple alarm-inventory resources at once
So far we were pushing to sysrepo each alarm-inventory change
separately. Unfortunately this is a computing intensive operations for
large data so we have to optimise it a bit.

This patch modified the class talking to sysrepo about hardware changes.
It now "caches" modifications of resources for an alarm and then pushes
them at once.

There is still a space for a small improvement: After this patch we
create one edit per alarm. This basically gets us with 4 calls (we have
4 alarms) for any change. I think this can be merged into one.
I don't think it is worth doing it now (the alarm-inventory does not
change that much for hardware related alarms). But of course, I can be
wrong.

I had to make some testing changes: all component paths were wrapped by
the COMPONENT macro inside another macro. Now we push more resources at
once and I don't know how to basically call a map function on
__VA_ARGS__ so I am wrapping all the components with COMPONENT manually
inside any macro call.

Change-Id: I9a77dde04e72011a25d6f6e7539fb9b81269e933