hardware: tweak the root component data
The description is not mandatory (and it's better set by cla-sysrepo,
anyway). The class attribute, on the other hand, is required, so let's
make sure that we produce valid data even when cla-sysrepo is not
running. (The operational DS doesn't *really* have to be valid, but hey,
this is super-cheap.)

Change-Id: Iabafbda427b9daec848ba6c7985e5397ad7f6746