hw: read static serial numbers from the ClearFog boards
On the most recent HW revisions of the SoM and the carrier board, there
should be some EEPROMs that are factory-populated with ONIE-compatible
TLV data. The EEPROMs are also semi-magic ones with some unique data at
the end of the memory space.

On some of these boards, the EEPROMs are not populated at all (which is
epxected on pre-1.3 "base" carrier and on some older SoM), so let's not
hard-fail when that happens. It's easily detectable by a missing
"eeprom" file in sysfs (but the dir entry for the I2C slave is always
present, as long as the Device Tree is not borked).

I have not seen an EEPROM with valid TLV data so far; the ones which we
have are usually empty. So I felt that the cleanest solution for this is
to put a new node just for the EEPROM into ietf-hardware, and put the
S/N from the unique part of each EEPROM in there. This required some
extra handling of the `oper-state` leaf so as not to confuse the
downstream consumers.

Change-Id: I36f810e4ca57b28405926da20a3c50bd520c480f