power: prevent spurious alarms about "missing devices" during init
When the background thread is "too slow" to initialize the I2C devices,
the main thread might run faster and request hwmon data before the
driver got a chance to bind to the device. When that happens, it looks
as if the HW device was missing (or that there's some I2C communication
error), like this:

 coherent-a-d-DQ000V9U ~ # velia-list-alarms
    Resource              Severity  Detail                                                            Last raised                          Status
 ✕  ne:pdu                critical  I2C read failure for PDU. Could not get hardware sensor details.  2024-10-10T19:15:23.292255297+00:00  cleared
 ✕  ne:psu1               critical  PSU is unplugged.                                                 2024-10-10T19:15:23.364300978+00:00  cleared
 ✕  ne:psu2               critical  PSU is unplugged.                                                 2024-10-10T19:15:23.446781485+00:00  cleared

This is really just a nuisance alarm, so make sure that it doesn't
happen during the initial service start.

Change-Id: Ia61dfe618ce1b44b49195d1eb272eee94a155bb6