hardware: don't update alarm-inventory in sideLoadedAlarms if not needed
The original code tried to update ietf-alarms:alarm-inventory every
time, even if the update was not necessary. We tried to add an entry
into ietf-alarms:alarm-inventory/alarm-type/resource leaf-list even when
we knew (or could have known) the entry was already there.

The solution is simple -- cache what we pushed into the list.

The original tests even assumed this happens -- see the diff. There is
an expectation that the update will happen at least once. It happened
in every invocation of the polling loop.
I modified the tests to check that we write into the alarm-inventory
only once.

Change-Id: I61685559ad18985092737bb5e0ab2b38585534a8