tests: simplify conversions to string
It feels silly to define printers for a vector/set/map of whatever; just
do the "whatever" once.

We should probably revisit this in future and look again at various
stringification options in doctest, but before that happens, well, let's
remove these duplicate things for STL containers.

(It's a pity that this requires implementing the printer for int64_t;
without that, a failure in tests/hardware_hwmon.cpp is not shown
properly.)

Change-Id: I8c1a51d63cd401bffae0007383f8d9abd8624633