hardware: workaround for wrong EEPROM product area length
Some of the EEPROMS in our PSUs are wrong. The content says that the
product area length will be 80 bytes but it is actually 88 bytes which
makes our parser unhappy and it rejects the input.

We have not seen (yet) any other errors than exactly this one.

This patch works around this bug by logging a warning and trying to
correct the length to 88 bytes in case this happens. In case we are
wrong about this correction, then we will *most likely* not successfully
parse anyway because the checksum will not be correct.

Change-Id: Ib5433c691c8f66d7a5d0e4f82f2ea9a626107ece