hardware: adjust voltage tolerances
On one board, the 3V3 channel was constantly measuring 3.406V. This was
below our threshold, so I tried to identify what the actual standard
tolerances are. In the ATX specification version 2.2 [1] (unofficial
link because the official (?) formfactors.org now redirects to an Intel
page which says nope, nothing here), all tolerances for positive
voltages are specified as ±5%.

Let's use that for the warning threshold everywhere (because this is a
standard, anything that's within the allowed range is by definition OK).

We also have to account for the tolerance of the actual measurement. The
PSU's specification says [2, page 27] that (during our typical load,
which fits into the 10%-20% range), the measurements have a ±5%
tolerance. So it's "possible" that we have, say, a perfectly valid
voltage of 5.25V (which is within 5% of the 5V target) measured by an
inaccurate sensor, yielding a 5.5125V. This shall be our "critical"
threshold.

TL;DR: a critical voltage alarm now means that the power is seriously
hosed. A warning might still be a false positive when the tolerances are
not aligned properly.

[1] https://cdn.instructables.com/ORIG/FS8/5ILB/GU59Z1AT/FS85ILBGU59Z1AT.pdf
[2] CzechLight/hw-docs, components/power/YH5151-1EB01R-Spec.pdf

Change-Id: Iad2e06ede62d82b5b338e8486224d85850d7c018