hardware: allow up to 50°C
The HW is specified [1, page 8] to work up to 50°C ambient temperature.
The PDU has three sensors and the PSU two sensors each. We have no idea
at which position they are placed, but it's fair to assume that none of
them is in a fridge. So, allow at least 50°C at any sensor (and we might
have to relax this even more in future). A simple device stack in a lab
reached 42°C with one PSU running.

[1] CzechLight/hw-docs, components/power/YH5151-1EB01R-Spec.pdf

Change-Id: If677bccdeb3f725ca85fc7da046c9a24e42971de