cli: velia-list-alarms support for status-history
Servers implementing ietf-alarms model with alarm-history provide
status-changes array that shows the history of the alarm. Let's show the
status history in our cli tool as well.

An example output might now look like this:

  ⏶  ne:psu1:voltage-5Vsb    critical  Sensor value crossed high threshold (11988 > 5400).     2024-10-16T15:53:17.455473252+00:00  active
  ⏶  ne:psu2:voltage-12V     cleared   Sensor value is within normal parameters.               2024-10-16T15:52:56.916417975+00:00  cleared
                             critical  Sensor value crossed high threshold (0 > 11300).        2024-10-16T15:44:42.659951986+00:00

The jq language is not very readable, I have tried to comment the code,
hopefully this will be enough.

Change-Id: Iebb865171ec67307a69295bae1c1e884a583b880