sysrepo: new approach to pushed ops data
There's been a change upstream, each session now has its own edit which
is used to build the final content of the operational DS. As a result,
we have to stop using the old approach with "discard" nodes. Instead,
there's now a "global" edit which is always re-applied upon any change.

Unfortunately, the new naming of the sysrepo-level API functions leave
something to be desired. We can now "discard" and "discard items", which
is a very different behavior yet a similar name. That's why the C++
wrapper tried to come up with better naming, and why we're propagating
these changes to the utility function here:

- "foreign removal" is about asking sysrepo to ensure that whatever
  stuff which ended up in the operational DS as a result of other
  session's stored ops data, or content that was propagated from the
  running DS, no matter what -- just ensure that it's blocked from
  propagation.

- "our removal" is about stuff that we might have pushed earlier, about
  ensuring that this gets undone.

Now, the funny thing happens when we are processing an async event from
netlink which informs us that an IP address is removed from an
interface. We cannot be sure whether this address was configured via
the running DS, or whether it's an address that was dynamically assigned
to an interface. We are also not sure whether we already processed an
earlier callback which might have notified us about this address being
assigned. When this happens, the simplest thing is to ensure that this
address is *really* not present in the operational DS, and therefore it
must be present in both "foreign removals" and "our removals". Not
pretty, but it works.

This currently creates duplicate entries in the /sysrepo:discard-items
opaque node, but that's just a sad part of life with older libyang that
will be, hopefully, fixed in some follow-up commit.

Change-Id: I0f48203a11539b3daae312825f6507ba64059a04
Depends-on: https://gerrit.cesnet.cz/c/CzechLight/dependencies/+/8134
Depends-on: https://gerrit.cesnet.cz/c/CzechLight/sysrepo-ietf-alarms/+/8137