eeprom: autodetect EEPROM type
This patch adds autodetection of EEPROM type so the users are not
required to specify it manually. The autodetection is actually very
dumb, we just try to read the file as if it was a known EEPROM type and
if it fails, we try the next one. Not very efficient but those files are
small so it should be fine.

Change-Id: Ice8e99e7ee7150ce45b1a8be1ac765c7a78b6eb4