utils: Fix valuesToYang creating wrong edits
We have seen on real devices that the /ietf-interfaces:interfaces
container in operation datastore was not present.

After some debugging we have discovered that sysrepo Session::editBatch
does not consider all the trees in the edit, but only those following
the passed edit tree (and it is actually properly documented[1]).

Our utils::valuesToYang function was creating a forest of trees, and the
handle DataNode was not necessarily the first sibling. This caused the
edit not to push the data we wanted.

This did not appear before, because (I think) we were always pushing
edit containing only one tree (i.e. one container of one module). But
after recent changes in sysrepo, the edit contains all the changes to
the operational DS done by this session, so now it is easy to have more
than one tree in the edit.
It is fascinating that the bug was there since, well, always.

[1] https://github.com/sysrepo/sysrepo/blob/ef93a1253cc97f13671759f6e7790cbf729a5ae9/src/sysrepo.h#L1023

Fixes: 98ad18de ("HardwareState: Add Sysrepo operational data callback")
Change-Id: I0106751f25c3983b92dc6b5a05ee48369c3bea2a