hw: let's call the USB-serial console port a "port"
The YANG model says that the `port` identity is used for "some sort of
networking port capable of receiving and/or transmitting networking
traffic". While the serial console is not your typical "network", I
still think that this makes better sense than calling it a `module`,
which refers to a "self-contained sub-system".

Change-Id: I7c3dea73ae02d6c076c8305e0822d8cd077e3148