ietf-hardware: propagate asset-id to the operational DS
...by activating a no-op subscriber in the running DS.

I'm not very happy about the validation. I wanted to offer some "nice"
diagnostics to the user, and to me, that translates to not being able to
set invalid/meaningless/ignored values. In context of the ietf-hardware
YANG module, this means that people should not be able to create
"strange" list entries for "invalid" hardware, and that miss-assigning
HW classes should be blocked as well.

At first I wanted to do this via some sysrepo callbacks -- just to make
sure that the system won't be announcing some validation rules which are
then disregarded in context of the operational DS. That did not work
well, though, because writes are allowed even when no subscribers are
active, and then we have that great fun of deciding what to do when a
SR_EV_ENABLED fails to apply (with an added bonus of not being able to
use some sysrepo APIs for error handling, such as generating
NETCONF-specific error messages). So, in the end I gave up and just
changed the schema. Hopefully the error message is clear enough...

The subscribing has to happen before the polling thread starts running,
obviously. Please don't ask me how I found out :].

Change-Id: I0d92248104a097da583c4462efd7019f367a7dec