network: ensure that all reconfigured interfaces are brought back up
The old code was taking all reconfigured interfaces down. The comment
said that `networkctl reload` was expected to bring them back up, but
nope, this was not the casse, at least not on inline amps with oscW and
oscE interfaces. What's puzzling is the fact that this works just one on
ROADM Line Degrees with the osc interface, but hey, life's too short to
debug crap like that for two days.

Change-Id: I47a4e1eb452b5c95b445e5d39f193ba4939dc871