system: propagage systemd's machine-id through NETCONF
We want to save machine-id [1] somewhere so users can read it.
The machine-id is system property, i.e., it should probably be saved
somwehere inside ietf-system module [2].

Initially I thought that ietf-system:system-state/platform/machine
leaf is the proper place to store the machine-id because the
description is "A vendor-specific identifier string representing the
hardware in use.". But then I realized that the node talks about
utsname [3], which is of course not machine-id.

So, I chose to augment the ietf-system:system-state/platform container
with a new leaf for machine-id.

[1] https://www.freedesktop.org/software/systemd/man/latest/machine-id.html
[2] https://datatracker.ietf.org/doc/rfc7317/
[3] https://www.man7.org/linux/man-pages/man2/uname.2.html

Change-Id: I301ff8c7edcaabc66158f7e9678ca61adedbb4f8