system: store local LLDP chassis id in sysrepo
Our devices use systemd's systemd-networkd for gathering and
transmitting LLDP data. For some time now, we already have our custom
YANG model (czechlight-lldp) that stores neighbor LLDP data in sysrepo.
If we look at those data, we would like to identify individual devices
using chassis id. So far, there is no way to get the chassis id from
sysrepo.

Systemd-networkd uses local machine id as chassis id. This is not
ideal, because the machine id should not be published in network, but
we have to wait until systemd-networkd supports some other way [1].

The point of this patch is to publish the machine id to sysrepo, so
that it can be fetched as chassis id in our czechlight-lldp model.
This patch adds container called 'local' into the model, where the
chassis id and chassis id type are stored as leafs.

[1] https://github.com/systemd/systemd/issues/37613

Change-Id: I301ff8c7edcaabc66158f7e9678ca61adedbb4f8