system: refactor network configuration
This patch follows what was proposed in [1], i.e.:
Now we always produce a .network file for systemd-networkd even if the
link is not defined in /ietf-interfaces:interfaces/interface list. When
this happens we generate an empty file into /run, which takes precedence
over .network file in /usr and thus effectively removes the default
configuration of the link.

However, this expects that the startup repository is filled with
something meaningful, otherwise we generate an empty file for every link
and thus the system won't be able to network.

[1] https://github.com/CESNET/velia/issues/12

Bug: https://github.com/CESNET/velia/issues/12
Change-Id: I0a1c82609c86e6d5c2463b3ba3435616844f1024