network: generate configs for all interfaces
We discussed in [1], that we should always produce a .network file for
systemd-networkd even if the link is not defined in
/ietf-interfaces:interfaces/interface list. When this happens we should
generate an empty file into /run, which takes precedence over .network
file in /usr and thus effectively removes the default
configuration of the link.

However, we would like to make this more robust in case there is some
kind of catchall .network configuration defined in the system which
would take precedence over our files.
Disabling such catchall file is possible by having a file that takes
priority over that file. Since one of the previous commits, we already
generate configurations named 10-<link>.network, which should take
precedence over the catchall file.
So the only remaning bit now is that if an user disables the interface
in sysrepo, we should generate a file that has a name that takes
precedence over any possible catchall file and disable the interface.

Beware that from now we expect that the running repository has
meaningful content at the moment velia-network starts.
If the interface is not defined, then velia generates a .network file that does not
"configure" the interface (autoconfiguration (DHCP, IPv6 RA or LLA) is
disabled and no IP addresses are listed).

This also means we want to make sure that at least one interface is
enabled and configured to make it hard accidentaly lose connectivity.
This is ensured by a new YANG must.

[1] https://github.com/CESNET/velia/issues/12
[2] https://www.freedesktop.org/software/systemd/man/latest/systemd.network.html

Depends-on:
Bug: https://github.com/CESNET/velia/issues/12
Change-Id: I0a1c82609c86e6d5c2463b3ba3435616844f1024