system: do not bring links down and up when reloading their configuration
In commit 08d59cd2f2 ("main: initialize ietf-interfaces config") we
implemented a reload of network configuration by taking the interface
down and then reloading the network configuration using networkctl
reload.
A comment in that commit explains that this is necessary because
systemd-networkd failed to perform some reconfigurations when a bridge
interface was removed. This was in 5/2021. Now, in 5/2025, things seem
to be going well so let's try to remove the workaround which apparently
caused some trouble on recent boxes and newer systemd, see e.g.
a854d2306a ("network: ensure that all reconfigured interfaces are
brought back up").

It may seem that the reload callback parameter, ChangedUnits, is now
unnecessary. In a manner of speaking, it is, but I think it can be
useful in testing.

See-also: 08d59cd2f2 ("main: initialize ietf-interfaces config")
See-also: a854d2306a ("network: ensure that all reconfigured interfaces are brought back up")
Change-Id: I2132dd86eaf0a41779127485dc502b3d72967282