tests: test YANG models using yanglint
In the czechlight-network YANG model we have quite a bunch of deviate
and must clauses that target ietf-interfaces model. Those were tested
using our C++ test codebase which made the file long and hard to read.
We realized that it would be much easier to test those using yanglint,
since yanglint tool is exactly for this.

This commit adds a framework for testing YANG models using yanglint
tool and migrates the tests for czechlight-network YANG model to it.

Change-Id: Ib06fe5bbf2de0e15b2af5d8e2287b6abec6e694b