network: get managed links from systemd-networkd
We had a hardcoded list of interfaces which are configurable and which
should be managed by velia. This list is hardcoded in velia code and
in czechlight-network YANG model.
This patch removes the hardcoded list from velia code, because the list
of interfaces can be fetched from systemd-networkd. We could do it via
systemd-networkd's D-Bus interface, or we can simply call networkctl
list with JSON output and parse it. This patch implements the later
option.

Note that the fetching is done only once in main, it does not react to
changes in the interface list so it can not react to situations like
if somebody plugs in a hotplug device (which is probably right anyway).

Change-Id: I7c62e682b2553edbfc86c7fbd7363fd01d2f798c