network: get managed links from systemd-networkd
We had a hardcoded list of interfaces which are configurable and which
should be managed by velia. This list is hardcoded in velia code and
in czechlight-network YANG model.
This patch removes the hardcoded list from velia code, because the list
of interfaces can be fetched from systemd-networkd using D-Bus. The
fetching is done in the factory function and only once and thus it
cannot react to hotpluggable devices (which is probably right anyway).

Change-Id: I7c62e682b2553edbfc86c7fbd7363fd01d2f798c