yang: do not enforce IP configuration for enabled protocol if the interface is disabled
In czechlight-network model we have some rules that make it hard for a
user to unintentionally misconfigure the network.
One of the rules said that if there is a protocol (IPv4 or IPv6) is
enabled on an interface, then there must be at least one IP assigned,
or the autococonfiguration is turned on.
However, I think that if the interface is disabled, there is no point in
enforcing this rule.

Change-Id: Idff561cb90b69fc2d28fda098742a8962109bb59