yang: new hardware specific models
During the redesign of the network configuration we noticed that our
current deviations do not account for different hardware types. We allow
a specified set of interfaces, but the set is not the same for all
hardware types.

It turned out this is not that easy to model in YANG. The
ietf-interfaces:interfaces/interface/name leaf is defined as a string
and we need to validate it against a set of allowed values.

The best approach (IMHO) would be to create a set of features for
individual interface types/names. Then we would create identities for
each interface name guarded by if-feature statement. Unfortunately, we
can't change the list key type.

Then we thought about deviations with if-feature, each definition would
be for an individual hardware type. Unfortunately, you can't use
if-feature in deviations/deviate context.

So, we ended up with a set of models for each hardware type. Each model
contains only one deviation and that is a must statement for interface
names.

All the hardware specific models import the base czechlight-network
model which define the common stuff and common must statements.
The specific model care only about interface names. All the specific
models also import the common model, but (and I was surprised) importing
does not mean that yanglint will pull the deviations from the common
model, so one has to add the common model explicitly as well.

Also, we added a revision date to the common and specific models and the
specific models import the common model with the new revision date. I
hope this will make obvious that those models are interconnected. Also,
when bumping czechlight-network version the YANG stack will probably
warn us that we have to check specific model deviations as well.

Change-Id: Ie84fd0a609c16ad2022a5571b5c6e1970e46ef60