doc: add systemd-networkd to the runtime dependencies
It was kinda-obvious from the preceding text already, but hey, we bother
with listing libsystemd and systemd, so list the netowrk part as well.
In fact, this is IMHO rather important because it's perfectly fine to
"use systemd" and still have something else, such as NetworkManager, for
network, but velia is picky here :).

Change-Id: I9c0e543da21f56778d9255879c44f1b61e8a29e4