network: generate config files with prefix 10-
This patch modifies the systemd-networkd configuration generator to use
a prefix of "10-" for the generated files. This is in line with the
common practice of using numerical prefixes to control the order of
processing for network configurations in systemd [1].
Files generated by velia are supposed to be preferred over any
default/catchall configuration files shipped by the system.

[1] https://www.freedesktop.org/software/systemd/man/latest/systemd.network.html

Change-Id: Ic5964fcc090dbba5c284fbfde0b916fc32f76862