network: static routes configuration
Until now, we used deviation for ietf-routing [1] model to make
the whole model config true and we used it only for reporting the
routing data. This patch enables configuring static routes for our SDN
devices using this model.
There are some simplifications though, as some of the model parts and
options are only for advanced routers (see [1, Appendix B]). We allow
only static routes with single next-hop interface, no direct routing is
supported yet.

The downside of this patch is that we have to use two sysrepo module
change callbacks. One for changes in the ietf-routing model and one for
changes in the ietf-interfaces model. This means that if user changes
something in both models, we execute the reload callback (in our case
`networkctl reload`) twice.

[1] https://datatracker.ietf.org/doc/html/rfc8349.html

Change-Id: Ie3932684eac90694c00db316b3c240ae7df0331a