network: static routes configuration
Until now, we used deviation for ietf-routing [1] model to make
the whole model config true and we used it only for reporting the
routing data. This patch enables configuring static routes for our SDN
devices using this model.
There are some simplifications though, as some of the model parts and
options are only for advanced routers (see [1, Appendix B]). We allow
setting only static routes with single next-hop interface, no special
next hops, multiple next hops, or manual setting of direct routing.
The RFC is a little bit confusing in this regard, as it says, quote:

> Every implementation of the core routing data model MUST provide
> exactly one instance of the "direct" pseudo-protocol type.

However, in the Appendix D of the same RFC, there is no direct protocol
instance in the example data tree. We decided to not require the direct
protocol instance.

The downside of this patch is that we have to use two sysrepo module
change callbacks. One for changes in the ietf-routing model and one for
changes in the ietf-interfaces model. This means that if user changes
something in both models, we execute the reload callback (in our case
`networkctl reload`) twice.

[1] https://datatracker.ietf.org/doc/html/rfc8349.html

Change-Id: Ie3932684eac90694c00db316b3c240ae7df0331a