refactor: make the ietf-hardware factory function more testable
A lot of this code is "only" parsing content from some sysfs entries,
which should be easily testable by providing some mock directory entries
for the unit tests. The only exception is code for power supplies which
relies on both /sys and /dev and some ioctls, so let's split this thing
into two parts.

Change-Id: Ia69613b7693104b595518d7a5130d9b32c1c949a